+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.0] - 2026-09-09
+
+### Changed
+
+- Upgraded `bevy` to `0.19`.
+- Updated the example to Bevy `0.19`.
+
+### Fixed
+
+- Feature gate compression helpers behind `writer` feature.
+
 ## [0.8.0] - 2026-01-13
 
 ### Changed
@@ -100,7 +111,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release.
 
-[Unreleased]: https://github.com/ldubos/bevy-histrion-packer/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/ldubos/bevy-histrion-packer/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/ldubos/bevy-histrion-packer/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/ldubos/bevy-histrion-packer/compare/v0.7.1...v0.8.0
 [0.7.1]: https://github.com/ldubos/bevy-histrion-packer/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/ldubos/bevy-histrion-packer/compare/v0.6.0...v0.7.0